Disentangle SpBLAS and SparseBLAS implementations
